+++ bindings/angular2/examples/alert-dialog.ts
@@ -3,73 +3,79 @@ import {
   Component,
   ComponentRef,
   AlertDialogFactory,
-  OnsAlertDialog,
   ViewChild,
-  AfterViewInit
+  ONS_DIRECTIVES,
+  Params,
+  OnInit,
+  OnDestroy
 } from '../src/angular2-onsenui';
 
 @Component({
   selector: 'div',
-  directives: [OnsAlertDialog],
+  directives: [ONS_DIRECTIVES],
   template: `
-    <ons-alert-dialog cancelable (cancel)="hide()">
+    <ons-alert-dialog cancelable #alert>
       <div class="alert-dialog-title">Warning!</div>
       <div class="alert-dialog-content">
-        This is just an example!
+        {{message}}
       </div>
       <div class="alert-dialog-footer">
-        <button class="alert-dialog-button" (click)="hide()">OK</button>
+        <button class="alert-dialog-button" (click)="alert.hide()">OK</button>
       </div>
     </ons-alert-dialog>
   `
 })
-class MyAlertDialogComponent implements AfterViewInit {
+class MyAlertDialogComponent {
+  message = '';
 
-  @ViewChild(OnsAlertDialog) private _alert: OnsAlertDialog;
-
-  constructor() {
-  }
-
-  ngAfterViewInit() {
-    this._alert.show();
-  }
-
-  hide() {
-    this._alert.hide();
+  constructor(params: Params) {
+    this.message = <string>params.at('message');
   }
 }
 
 @Component({
   selector: 'app',
   providers: [AlertDialogFactory],
+  directives: [ONS_DIRECTIVES],
   template: `
   <ons-page class="page">
     <ons-toolbar>
       <div class="center">Alert Dialog</div>
     </ons-toolbar>
-    <div class="page__background"></div>
-    <div class="page__content">
+    <div class="content">
       <div style="text-align: center; margin: 10px">
-        <ons-button (click)="showAlertDialog()">create Alert Dialog</ons-button>
+        <ons-button (click)="showAlertDialog()">show</ons-button>
       </div>
     </div>
   </ons-page>
   `
 })
-export class AppComponent implements AfterViewInit {
-  private _alert: ComponentRef<MyAlertDialogComponent>;
+export class AppComponent implements OnInit, OnDestroy {
+  private _alert: any;
+  private _destroyAlert: Function;
 
-  constructor(private adf: AlertDialogFactory) {
+  constructor(private _adf: AlertDialogFactory) {
   }
 
-  ngAfterViewInit() {
+  ngOnInit() {
+    this._adf
+      .createAlertDialog(MyAlertDialogComponent, {message: 'This is just an example.'})
+      .then(({alertDialog, destroy}) => {
+        this._alert = alertDialog;
+        this._destroyAlert = destroy;
+      });
   }
 
   showAlertDialog() {
-    this.adf.createAlertDialog(MyAlertDialogComponent).then(componentRef => {
-      this._alert = componentRef;
-    });
+    if (this._alert) {
+      this._alert.show();
+    }
+  }
+
+  ngOnDestroy() {
+    this._destroyAlert();
   }
 }
 
 bootstrap(AppComponent);
+